Feat: Add missing charts for v2.1 (#289)
* refactor(v2.1): rename AreaChart to SimpleAreaChart

* feat(v2.1): make chart tooltip show only defined labels

* feat(v2.1): split two values chart logic into a reusable component

Also, add a possibility of having just the primary value of the chart

* feat(v2.1): add AreaChart based on two values
