+++ frontend/src/components/charts/AreaChart.tsx
@@ -1,53 +1,46 @@
-import { AreaChart as RechartsArea, Area, ResponsiveContainer } from "recharts";
+import { Area } from "recharts";
 
-type Props = {
-  /**
-   * The data to be displayed in the chart.
-   */
-  data: {
-    name: string;
-    value: number;
-  }[];
-};
+import { BaseTwoValuesChart } from "./BaseTwoValuesChart";
+import { BaseTwoValuesChartProps } from "./BaseTwoValuesChart/types";
+
+type AreaChartProps = Omit<BaseTwoValuesChartProps, "xContentPaddingEnabled">;
 
 /**
- * Area chart component. It is used to display a chart with a series-based data.
- * The series should be an array of objects with the following properties:
+ * Area Chart component. It is used to display a chart with a series-based data.
  *
- * - name: string
- * - value: number
- *
- * This chart does not render X-Axis, Y-Axis, Legend, or Tooltip.
- *
- * The width and height of the chart are based on the parent container, and the chart take 100% of the space.
+ * @see {@link BaseTwoValuesChart} for more details on how this data and props work.
  */
-export const AreaChart = ({ data }: Props) => {
+export const AreaChart = ({
+  primaryValueName,
+  secondaryValueName,
+  primaryValueOnly = false,
+  ...props
+}: AreaChartProps) => {
   return (
-    <ResponsiveContainer width="100%" height="100%">
-      <RechartsArea data={data}>
-        <defs>
-          <linearGradient id="colorGradient" x1="0" y1="0" x2="0" y2="1">
-            <stop
-              offset="80%"
-              stopColor="var(--pal-graph-gradient-primary)"
-              stopOpacity={0.1}
-            />
-            <stop
-              offset="100%"
-              stopColor="var(--pal-graph-gradient-secondary)"
-              stopOpacity={0}
-            />
-          </linearGradient>
-        </defs>
+    <BaseTwoValuesChart primaryValueOnly={primaryValueOnly} {...props}>
+      <Area
+        type="monotone"
+        dataKey="primaryValue"
+        stroke="var(--pal-graph-primary)"
+        fill="var(--pal-graph-primary)"
+        name={primaryValueName}
+        strokeWidth={2}
+      />
+
+      {!primaryValueOnly && (
         <Area
-          stroke="var(--pal-graph-primary)"
-          strokeOpacity={0.6}
-          strokeWidth={1.5}
           type="monotone"
-          dataKey="value"
-          fill="url(#colorGradient)"
+          dataKey="secondaryValue"
+          stroke="var(--pal-graph-secondary)"
+          fill="var(--pal-graph-secondary)"
+          name={secondaryValueName}
+          strokeWidth={2}
         />
-      </RechartsArea>
-    </ResponsiveContainer>
+      )}
+    </BaseTwoValuesChart>
   );
 };
+
+AreaChart.TimeRange = BaseTwoValuesChart.TimeRange;
+
+export const TimeRange = AreaChart.TimeRange;
